fix: Fold InternalUpdate behavior into base class implementation.
