+++ Assets/Scripts/StandardSamples/UI/Peer2Peer/HighFrequency/UIHighFrequencyPeer2PeerMenu.cs
@@ -95,7 +95,7 @@ namespace PlayEveryWare.EpicOnlineServices.Samples
                 Peer2PeerManager.sendActive = !Peer2PeerManager.sendActive;
             }
         }
-        private void Update()
+        protected override void InternalUpdate()
         {
             ProductUserId messageFromPlayer = Peer2PeerManager.HandleReceivedMessages();
             if (messageFromPlayer != null)
